create transaction module, move transaction builder there
