+++ rust/tw_bitcoin/src/input/mod.rs
@@ -1,5 +1,4 @@
-use crate::InputContext;
-use bitcoin::{ScriptBuf, TxIn, Witness};
+use bitcoin::{ScriptBuf, TxIn, TxOut, OutPoint, Sequence, Witness};
 
 mod p2pkh;
 mod p2tr_key_path;
@@ -10,6 +9,32 @@ pub use p2pkh::*;
 pub use p2tr_key_path::*;
 pub use p2tr_script_path::*;
 pub use p2wpkh::*;
+
+#[derive(Debug, Clone)]
+pub struct InputContext {
+    pub previous_output: OutPoint,
+    pub value: u64,
+    // The condition for claiming the output.
+    pub script_pubkey: ScriptBuf,
+    pub sequence: Sequence,
+    // Witness data for Segwit/Taproot transactions.
+    // TODO: Remove this?
+    pub witness: Witness,
+}
+
+impl InputContext {
+    pub fn new(utxo: TxOut, point: OutPoint) -> Self {
+        InputContext {
+            previous_output: point,
+            value: utxo.value,
+            script_pubkey: utxo.script_pubkey,
+            // Default value of `0xFFFFFFFF = 4294967295`.
+            sequence: Sequence::default(),
+            // Empty witness.
+            witness: Witness::new(),
+        }
+    }
+}
 
 #[derive(Debug, Clone)]
 pub enum TxInput {
